+++ packages/messaging/ios/Plugin/FirebaseMessagingHelper.swift
@@ -1,0 +1,15 @@
+import Foundation
+import Capacitor
+import FirebaseCore
+
+public class FirebaseMessagingHelper {
+    public static func createNotificationResult(notification: UNNotification) -> JSObject {
+        var result = JSObject()
+        result["body"] = notification.request.content.body
+        result["data"] = JSTypes.coerceDictionaryToJSObject(notification.request.content.userInfo) ?? [:]
+        result["id"] = notification.request.identifier
+        result["subtitle"] = notification.request.content.subtitle
+        result["title"] = notification.request.content.title
+        return result
+    }
+}